refactor(utils): split oversized codegraph helpers
